백준 0921

+++ rectangle_and_triangle/src/main/java/com/baekjoon/ex_15894.java
@@ -2,7 +2,7 @@ package com.baekjoon;
 
 public class ex_15894 {
     public static void main(String[] args) {
-        int n = new java.util.Scanner(System.in).nextInt();
+        long n = new java.util.Scanner(System.in).nextInt();
         System.out.println(4*n);
     }
 }
